feat: export prettyShadowDOM and logShadowDOM
