+++ src/debug.ts
@@ -1,10 +1,11 @@
 import { screen } from "@testing-library/dom";
-import { prettyShadowDOM } from "../src/prettyShadowDOM";
+import { logShadowDOM } from "./log-shadow-dom";
+import { prettyShadowDOM } from "./pretty-shadow-dom";
 
 export function debug(...args: Parameters<typeof screen.debug>) {
   let [element, maxLength, options] = args;
 
   Array.isArray(element)
-    ? element.forEach((el) => prettyShadowDOM(el, maxLength, options))
-    : prettyShadowDOM(element, maxLength, options);
+    ? element.forEach((el) => logShadowDOM(el, maxLength, options))
+    : logShadowDOM(element, maxLength, options);
 }
