+++ tsconfig.json
@@ -10,7 +10,6 @@
 
     "strict": true,
     "esModuleInterop": true,
-    "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
     "moduleResolution": "node",
     "jsx": "react"
